Extract Embeddings Run Collection Logic To Service
